Update to block spacing system.
Original block spacing system allowed for 1 minute blocks for each
Proof-Of system. Because we use both POW and POS permanently, the result
was two blocks per minute.

This adjusts the target spacing to take into account the number of Proof
systems in use.

It also sets this change to take effect after block 86400, or in about
29 days.