Updated Message System
Now called 'Private Messages' instead of 'Encrypted Messages'.
Receiving addresses now have a context menu entry to 'Copy PM Key' which
will copy everything needed to send them a PM to the clipboard. Just
paste the information in an email, chat, or any other text window and
send it to your friend.

Also fixed chat text showing up in as white by default. Removed the
colored text highlighting - now the color selector changes the color of
the text and not the whitespace around the text.

Everything else seems to work fine. Still needs some ease of use
updating and the ability to update labels for people not in your address
book, but it's still leagues better than what it was.